// src/app/shared/types/action-type.enum.ts
export enum ActionType {
  ADD_USER_TO_GAME = 'addUserToGame',
  CHAT = 'chat',
  CLEAR_CANVAS = 'clearCanvas',
  DRAW = 'draw',
  DRAWER_WORD = 'drawerWord',
  GAME_ENDED_RANKING = 'gameEndedRanking',
  NEW_ROUND = 'newRound',
  USER_JOINED = 'userJoined',
  ROUND_STATISTICS = 'roundStatistics',
  START_GAME = 'startGame'
}